Refactor package name: fabric -> internal/fabric
