+++ internal/fabric/core/comm/config.go
