+++ internal/fabric/core/comm/util.go
