+++ internal/fabric/protoutil/commonutils.go
