+++ internal/fabric/protoutil/proputils.go
