+++ pkg/infra/crypto.go
@@ -11,8 +11,8 @@ import (
 	"io/ioutil"
 	"math/big"
 
-	"tape/fabric/bccsp/utils"
-	"tape/fabric/common/crypto"
+	"tape/internal/fabric/bccsp/utils"
+	"tape/internal/fabric/common/crypto"
 
 	"github.com/hyperledger/fabric-protos-go/common"
 	"github.com/pkg/errors"
